refac: 재사용성 향상을 위해 readImageMetadata파일을 util 파일로 분리
