+++ src/hooks/useImageMetadata.ts
@@ -1,54 +1,13 @@
 import { useEffect, useRef, useState } from "react";
-import { validateImageFile, type Limit } from "../utils/validate";
+import { type Limit } from "../utils/validate";
+import {
+  type ImageFileMetadata,
+  readImageMetadata,
+} from "../utils/readImageMetadata";
 
 interface Props {
   limit?: Limit;
 }
-
-export interface ImageFileMetadata {
-  height: number;
-  width: number;
-  size: number;
-  name: string;
-  extension: string;
-  src: string;
-}
-
-export const readImageFileMetadata = (
-  file: File,
-  limit?: Limit
-): Promise<ImageFileMetadata | null> => {
-  return new Promise((resolve) => {
-    const reader = new FileReader();
-    reader.onload = (e) => {
-      const readerTarget = e.target as FileReader;
-      const image = new Image();
-      image.src = readerTarget.result as string;
-
-      image.onload = () => {
-        const metaData: ImageFileMetadata = {
-          height: image.height,
-          width: image.width,
-          size: file.size,
-          name: file.name,
-          extension: (file.name.split(".").pop() ?? "").toLowerCase(),
-          src: image.src,
-        };
-
-        if (limit) {
-          const validatePassed = validateImageFile(limit, metaData);
-          if (!validatePassed) return resolve(null);
-        }
-        resolve(metaData);
-      };
-
-      image.onerror = () => resolve(null);
-    };
-
-    reader.onerror = () => resolve(null);
-    reader.readAsDataURL(file);
-  });
-};
 
 export const checkFileType = (inputEl: HTMLInputElement) => {
   if (inputEl.type === "file") return true;
@@ -70,7 +29,7 @@ const useImageMetadata = ({ limit }: Props | undefined = {}) => {
     const file = target.files?.[0];
 
     if (file) {
-      const metaData = await readImageFileMetadata(file, limit);
+      const metaData = await readImageMetadata(file, limit);
       setImageMetadata(metaData);
     }
   };
